stdint compatibility

+++ libsrng.c
@@ -8,13 +8,13 @@ struct libsrng_stable_random_state {
   unsigned linear:  8;
 };
 
-static inline unsigned short libsrng_random_linear(unsigned short);
-static inline unsigned char libsrng_random_combined(unsigned long long *);
-static inline unsigned long long libsrng_random_combined_multibyte(unsigned long long *, unsigned char);
-static inline unsigned short libsrng_random_halfword(unsigned long long *);
-static inline unsigned long long libsrng_random_seed(unsigned long long *);
+static inline uint16_t libsrng_random_linear(uint16_t);
+static inline unsigned char libsrng_random_combined(uint64_t *);
+static inline uint64_t libsrng_random_combined_multibyte(uint64_t *, unsigned char);
+static inline uint16_t libsrng_random_halfword(uint64_t *);
+static inline uint64_t libsrng_random_seed(uint64_t *);
 static inline unsigned char libsrng_stable_random(struct libsrng_stable_random_state *);
-static inline unsigned short libsrng_random_range(unsigned long long *, unsigned short);
+static inline uint16_t libsrng_random_range(uint64_t *, uint16_t);
 
 #define HALFWORD_LCG_MULTIPLIER             0x6329
 #define HALFWORD_LCG_ADDEND                 0x4321
@@ -24,40 +24,40 @@ static inline unsigned short libsrng_random_range(unsigned long long *, unsigned
 
 #define STABLE_RANDOM_NEXT_LINEAR(s) ((s) -> linear *= 73, (s) -> linear += 29, (s) -> linear)
 
-unsigned short libsrng_random (unsigned long long * state, unsigned short range, unsigned reseed) {
+uint16_t libsrng_random (uint64_t * state, uint16_t range, unsigned reseed) {
   if (!state) return 0;
   while (reseed --) *state = libsrng_random_seed(state);
   return libsrng_random_range(state, range);
 }
 
-static inline unsigned short libsrng_random_linear (unsigned short previous) {
+static inline uint16_t libsrng_random_linear (uint16_t previous) {
   return (previous * HALFWORD_LCG_MULTIPLIER + HALFWORD_LCG_ADDEND) & 0xffff;
 }
 
-static inline unsigned char libsrng_random_combined (unsigned long long * state) {
+static inline unsigned char libsrng_random_combined (uint64_t * state) {
   // this conditional should be determined at compile time, and it should be true for virtually any reasonable platform
   if (sizeof *state == sizeof(struct libsrng_stable_random_state)) {
-    union {struct libsrng_stable_random_state s; unsigned long long n;} * state_union = (void *) state;
+    union {struct libsrng_stable_random_state s; uint64_t n;} * state_union = (void *) state;
     return libsrng_stable_random(&(state_union -> s));
   } else {
     struct libsrng_stable_random_state temp_state = {.shift = *state, .carry = *state >> 32, .current = *state >> 40,
                                                      .prev = *state >> 48, .linear = *state >> 56};
     unsigned char result = libsrng_stable_random(&temp_state);
-    *state = ((unsigned long long) temp_state.shift) | ((unsigned long long) temp_state.carry << 32) | ((unsigned long long) temp_state.current << 40) |
-             ((unsigned long long) temp_state.prev << 48) | ((unsigned long long) temp_state.linear << 56);
+    *state = ((uint64_t) temp_state.shift) | ((uint64_t) temp_state.carry << 32) | ((uint64_t) temp_state.current << 40) |
+             ((uint64_t) temp_state.prev << 48) | ((uint64_t) temp_state.linear << 56);
     return result;
   }
 }
 
-static inline unsigned long long libsrng_random_combined_multibyte (unsigned long long * state, unsigned char width) {
-  if (width > sizeof(unsigned long long)) return -1;
-  unsigned long long result = 0;
+static inline uint64_t libsrng_random_combined_multibyte (uint64_t * state, unsigned char width) {
+  if (width > sizeof(uint64_t)) return -1;
+  uint64_t result = 0;
   while (width --) result = (result << 8) | libsrng_random_combined(state);
   return result;
 }
 
-static inline unsigned short libsrng_random_halfword (unsigned long long * state) {
-  unsigned short buffer = libsrng_random_combined_multibyte(state, 2);
+static inline uint16_t libsrng_random_halfword (uint64_t * state) {
+  uint16_t buffer = libsrng_random_combined_multibyte(state, 2);
   unsigned char count = libsrng_random_combined(state);
   unsigned char shift = count >> 4, multiplier = 3 + ((count & 12) >> 1);
   count = (count & 3) + 2;
@@ -66,10 +66,10 @@ static inline unsigned short libsrng_random_halfword (unsigned long long * state
   return (buffer * multiplier) & 0xffff;
 }
 
-static inline unsigned long long libsrng_random_seed (unsigned long long * state) {
-  unsigned long long first = libsrng_random_combined_multibyte(state, 8);
+static inline uint64_t libsrng_random_seed (uint64_t * state) {
+  uint64_t first = libsrng_random_combined_multibyte(state, 8);
   first = first * SEED_LCG_MULTIPLIER + SEED_LCG_FIRST_ADDEND;
-  unsigned long long second = 0;
+  uint64_t second = 0;
   unsigned count;
   for (count = 0; count < 4; count ++) second = (second << 16) + libsrng_random_halfword(state);
   second = second * SEED_LCG_MULTIPLIER + SEED_LCG_SECOND_ADDEND;
@@ -77,7 +77,7 @@ static inline unsigned long long libsrng_random_seed (unsigned long long * state
 }
 
 static inline unsigned char libsrng_stable_random (struct libsrng_stable_random_state * state) {
-  unsigned p;
+  uint32_t p;
   if (!state -> shift) for (p = 0; p < 4; p ++) state -> shift = (state -> shift << 8) | STABLE_RANDOM_NEXT_LINEAR(state);
   state -> shift ^= state -> shift >> 8;
   state -> shift ^= state -> shift << 9;
@@ -107,12 +107,12 @@ static inline unsigned char libsrng_stable_random (struct libsrng_stable_random_
   }
 }
 
-static inline unsigned short libsrng_random_range (unsigned long long * state, unsigned short limit) {
+static inline uint16_t libsrng_random_range (uint64_t * state, uint16_t limit) {
   if (limit == 1) return 0;
-  unsigned short result = libsrng_random_halfword(state);
+  uint16_t result = libsrng_random_halfword(state);
   if (!(limit & (limit - 1))) return result & (limit - 1);
   if (result >= limit) return result % limit;
-  unsigned short resampling_limit = 0x10000 % limit;
+  uint16_t resampling_limit = 0x10000 % limit;
   while (result < resampling_limit) result = libsrng_random_halfword(state);
   return result % limit;
 }
