+++ libsrng.h
@@ -2,6 +2,8 @@
 
 #define ___LIBSRNG
 
+#include <stdint.h>
+
 #ifdef __cplusplus
   // only one declaration, so no braces
   extern "C"
@@ -13,6 +15,6 @@
 //         generate a value at all - this can be used to reseed a state without consuming a random number.
 // reseed: used to generate a new seed for the state; if non-zero, the state will be reseeded this number of times
 //         (using a different, and slower, RNG) before generating a random number. This allows for multiple sequences.
-unsigned short libsrng_random(unsigned long long * state, unsigned short range, unsigned reseed);
+uint16_t libsrng_random(uint64_t * state, uint16_t range, unsigned reseed);
 
 #endif
